// types/schemas-to-ts/api/product-brand.ts
// Interface automatically generated by schemas-to-ts

export interface ProductBrand {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    name: string;
    description?: string;
  };
}
export interface ProductBrand_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  name: string;
  description?: string;
}

export interface ProductBrand_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  name: string;
  description?: string;
}

export interface ProductBrand_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  name: string;
  description?: string;
}